+++ CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tokio: Add `task::Builder::new_current_thread` but panic inside.
 - tonic: Add `service` module and `Extensions`.
 - tonic: Support interceptor.
+- etcd: Support load and dump in toml format.
 
 ### Fixed
 
 - tonic: Fix passing metadata in request and response. Add `content-type` and `date` field.
 - tonic: Fix panic on unimplemented error.
+- etcd: Fix lease grant.
 
 ## [0.2.10] - 2022-11-09